feat(structure): branching the rendering of different document list panes
