+++ packages/sanity/src/core/components/previews/types.ts
@@ -13,7 +13,7 @@ export type PortableTextPreviewLayoutKey = 'block' | 'blockImage' | 'inline'
  *
  * @public
  */
-export type GeneralPreviewLayoutKey = 'compact' | 'default' | 'media' | 'detail'
+export type GeneralPreviewLayoutKey = 'compact' | 'default' | 'media' | 'detail' | 'sheetList'
 
 /**
  * Preview layout key. See also {@link GeneralPreviewLayoutKey} and {@link PortableTextPreviewLayoutKey}
