+++ packages/sanity/src/core/preview/components/_previewComponents.ts
@@ -14,7 +14,6 @@ export const _previewComponents = {
   compact: CompactPreview,
   default: DefaultPreview,
   detail: DetailPreview,
-  sheetList: DefaultPreview,
   inline: InlinePreview,
   media: MediaPreview,
 } as const
