+++ packages/sanity/src/structure/panes/documentList/DocumentSheetListPane.tsx
@@ -1,0 +1,8 @@
+import {Text} from '@sanity/ui'
+
+function DocumentSheetListPane() {
+  // eslint-disable-next-line i18next/no-literal-string
+  return <Text>SheetList view</Text>
+}
+
+export default DocumentSheetListPane
